feat: first pass as making subject chips clickable
