--- solution/ui/regulations/eregs-component-lib/src/components/shared-components/results-item-parts/SubjectChip.vue
@@ -1,12 +1,0 @@
-<script setup>
-const props = defineProps({
-    subjectName: {
-        type: String,
-        required: true,
-    },
-});
-</script>
-
-<template>
-    <div class="subject__chip">{{ subjectName }}</div>
-</template>
