+++ solution/ui/regulations/eregs-vite/src/components/policy-repository/SubjectChip.vue
@@ -1,0 +1,24 @@
+<script setup>
+const props = defineProps({
+    subjectId: {
+        type: Number,
+        required: true,
+    },
+    subjectName: {
+        type: String,
+        required: true,
+    },
+});
+</script>
+
+<template>
+    <router-link
+        class="subject__chip"
+        :to="{
+            name: 'policy-repository',
+            query: { subjects: [subjectId.toString()] },
+        }"
+    >
+        {{ subjectName }}
+    </router-link>
+</template>
